fixed bug overflow sub

+++ src/lib.rs
@@ -31,8 +31,8 @@ pub fn to_fixed(x: f64, m: u8, n: u8) -> Result<(u64, bool), String> {
                               // is where _you_ interpret the point to be, which depends on `exp` at this point.
                               // now all you have to do is slice out the fractional and non-fractional parts individually.
 
-    let fractional_part = bits & mask(MANT_SIZE - exp as u64);
-    let integer_part = bits >> (MANT_SIZE - exp as u64);
+    let fractional_part = bits & mask((MANT_SIZE as i32 - exp as i32) as u64);
+    let integer_part = bits >> ((MANT_SIZE as i32 - exp as i32) as u64);
 
     // now, depending on `m` and `n` you need to figure out whether rouding occurs.
     // if that's the case, that information is reported back to the user via the `is_exact` flag.
@@ -41,20 +41,20 @@ pub fn to_fixed(x: f64, m: u8, n: u8) -> Result<(u64, bool), String> {
     let integer_part_on_m_bits = integer_part & mask(m as u64);
 
     let mut fractional_part_on_n_bits = (fractional_part
-        & (mask(n as u64) << (MANT_SIZE - exp as u64 - n as u64)))
-        >> (MANT_SIZE - exp as u64 - n as u64);
+        & (mask(n as u64) << ((MANT_SIZE as i32 - exp as i32) as u64 - n as u64)))
+        >> ((MANT_SIZE as i32 - exp as i32) as u64 - n as u64);
 
     if integer_part_on_m_bits < integer_part {
         return Err("Integer field does not fit into `m`.".to_string());
     }
 
-    let round_bit = fractional_part >> (MANT_SIZE - exp as u64 - (n as u64 + 1)) & 1 != 0;
+    let round_bit = fractional_part >> ((MANT_SIZE as i32 - exp as i32) as u64 - (n as u64 + 1)) & 1 != 0;
 
     if round_bit {
         fractional_part_on_n_bits += 1;
     }
 
-    let sticky_bit = fractional_part & mask(MANT_SIZE - exp as u64 - (n as u64 + 1)) != 0;
+    let sticky_bit = fractional_part & mask((MANT_SIZE as i32 - exp as i32) as u64 - (n as u64 + 1)) != 0;
 
     let is_exact = !sticky_bit && !round_bit;
     let ans = (integer_part_on_m_bits << n) + fractional_part_on_n_bits;
